+++ services/sharepoint-connector/src/sharepoint-synchronization/sharepoint-synchronization.module.ts
@@ -1,15 +1,21 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { MicrosoftApisModule } from '../microsoft-apis/microsoft-apis.module';
+import { PermissionsSyncModule } from '../permissions-sync/permissions-sync.module';
 import { ProcessingPipelineModule } from '../processing-pipeline/processing-pipeline.module';
 import { UniqueApiModule } from '../unique-api/unique-api.module';
 import { ContentSyncService } from './content-sync.service';
-import { PermissionsSyncService } from './permissions-sync.service';
 import { SharepointSynchronizationService } from './sharepoint-synchronization.service';
 
 @Module({
-  imports: [ConfigModule, MicrosoftApisModule, UniqueApiModule, ProcessingPipelineModule],
-  providers: [SharepointSynchronizationService, ContentSyncService, PermissionsSyncService],
+  imports: [
+    ConfigModule,
+    MicrosoftApisModule,
+    UniqueApiModule,
+    ProcessingPipelineModule,
+    PermissionsSyncModule,
+  ],
+  providers: [SharepointSynchronizationService, ContentSyncService],
   exports: [SharepointSynchronizationService],
 })
 export class SharepointSynchronizationModule {}