+++ services/sharepoint-connector/src/unique-api/unique-users/unique-users.consts.ts
@@ -1,0 +1,30 @@
+import { gql } from 'graphql-request';
+
+export interface ListUsersQueryInput {
+  skip: number;
+  take: number;
+}
+
+export interface ListUsersQueryResult {
+  listUsers: {
+    totalCount: number;
+    nodes: {
+      id: string;
+      active: boolean;
+      email: string;
+    }[];
+  };
+}
+
+export const LIST_USERS_QUERY = gql`
+  query ListUsers($skip: Int!, $take: Int!) {
+    listUsers: paginatedUsers(skip: $skip, take: $take) {
+      totalCount
+      nodes {
+        id
+        active
+        email
+      }
+    }
+  }
+`;